Move the Gnomish Mocha bridge up a directory level
